backend/requirements.txt
Flask==3.0.2
flask-cors==4.0.0
tensorflow>=2.13.0
Pillow>=10.0.0
bcrypt==4.1.2
PyJWT==2.8.0
openpyxl==3.1.2
numpy>=1.24.3
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0
